Module Libabsolute

module Csp : sig ... end

module Print : sig ... end

Printing functions for the different types in Csp

module Parser : sig ... end

module Signature : sig ... end

Module for Abstract Domains for Constraint Programming (ADCP) they must feature consistency, split and precision operators.

module Iterator : sig ... end

module Solver : sig ... end

module Result : sig ... end

module Constant : sig ... end

module Domains : sig ... end

This module handle the domain environment, already equipped with some basic domains. You can add your own using the register_* functions, and combine them with the already defined ones.

module Kleene : sig ... end

module Consistency : sig ... end

This module provides types and operations for handling consistencies. A consitency is a property obtained after a filtering operation f(s,p): given an abstract value s, and a predicate p, it computes a set s' \subseteq s such that : \forall x \in s, p(x) \implies x \in s'