Module Domains.Product

Specialized Reduced Product. Corresponds to the option -d product (a,b) of AbSolute. if a constraint can be filtered exactly by the domain b, it is affected to it, otherwise it is affected to b.

module Make (A : Signature.Domain) (B : Signature.Domain) : Signature.Domain