Module Libabsolute.Polynom

Module for polynomials over an abstract arithmetic rings

This modules handles symbolic computations over multivariate polynoms It is parametrized by a ring module which deals with basic arithmetic *

module Make (R : Ring.T) : sig ... end

module Int : sig ... end

module Float : sig ... end

module Rational : sig ... end