Module Libabsolute.Result

This module defines solution of the abstract solver as covers

type 'a t = {

sure : 'a list;
(*
elements that satisfy the constraints
*)

unsure : 'a list;
(*
elements that MAY satisfy the constraints
*)

nb_sure : int;
(*
size of sure list
*)

nb_unsure : int;
(*
size of unsure list
*)

nb_steps : int;
(*
number of steps of the solving process
*)

best_value : Q.t;
(*
best value found during the optimization
*)

}

The type of cover where sure is an under-approximation of the solution set and the union of sure and unsure is an over-approximaton of the solution set

val empty : 'a t

empty result

val add_inner : 'a t -> 'b -> 'c t

adds an inner element to a result

val add_outer : 'a t -> 'b -> 'c t

adds an outer element to a result

val print : Stdlib.Format.formatter -> 'a t -> unit

printer