Module Apron_domain.MAKE

Parameters

AP : ADomain

Signature

module A = Apronext.Abstractext

module E = Apronext.Environmentext

type t = AP.t A.t

include sig ... end

type internal_constr = Apron.Tcons1.t

val man : AP.t Apron.Manager.t

val of_expr : Apron.Environment.t -> Libabsolute.Expr.t -> Apron.Texpr1.t

val of_cmp_expr : 'a Apron.Abstract1.t -> (Libabsolute.Expr.t * Libabsolute.Constraint.cmpop * Libabsolute.Expr.t) -> Apronext.Tconsext.t

val apron_to_var : 'a Apron.Abstract1.t -> string list * string list

val to_expr : Apron.Texpr1.expr -> Libabsolute.Expr.t

val apron_to_bexpr : Apron.Tcons1.t -> Libabsolute.Expr.t * Libabsolute.Constraint.cmpop * Libabsolute.Expr.t

val to_constraint : AP.t Apronext.Abstractext.t -> Libabsolute.Constraint.t

val empty : AP.t A.t

val internalize : ?⁠elem:'a Apron.Abstract1.t -> (Libabsolute.Expr.t * Libabsolute.Constraint.cmpop * Libabsolute.Expr.t) -> Apronext.Tconsext.t

val externalize : Apron.Tcons1.t -> Libabsolute.Expr.t * Libabsolute.Constraint.cmpop * Libabsolute.Expr.t

val vars : 'a A.t -> (Libabsolute.Csp.typ * string) list

val dom_to_texpr : Apron.Environment.t -> Libabsolute.Csp.dom -> Apron.Texpr1.t

val add_var : AP.t A.t -> (Libabsolute.Csp.typ * string * Libabsolute.Csp.dom) -> AP.t A.t

val var_bounds : AP.t A.t -> string -> Mpqf.t * Mpqf.t

val rm_var : AP.t A.t -> string -> AP.t A.t

val is_empty : AP.t A.t -> bool

val join : AP.t A.t -> AP.t A.t -> AP.t A.t * bool

val meet : AP.t A.t -> AP.t A.t -> AP.t A.t option

val diff : 'a option

val filter : AP.t A.t -> Apronext__.Tconsext.t -> AP.t A.t Libabsolute.Consistency.t

val print : Stdlib.Format.formatter -> 'a A.t -> unit

val pman : Polka.strict Polka.t Apron.Manager.t

val to_poly : AP.t A.t -> Apron.Environment.t -> Polka.strict Polka.t A.t

computes the smallest enclosing polyhedron

val forward_eval : AP.t A.t -> Libabsolute.Expr.t -> Mpqf.t * Mpqf.t

interval evaluation of an expression within an abtract domain

val largest : AP.t A.t -> Apron.Var.t * Apron.Interval.t * Libabsolute.Q.t

val minmax : Apronext.Intervalext.t array -> int -> Mpqf.t -> int -> Mpqf.t -> Apron.Scalar.t * int * Apron.Scalar.t

val generate_linexpr : E.t -> float array -> float array -> (Apron.Coeff.t * Apron.Var.t) list * (Apron.Coeff.t * Apron.Var.t) list * float

val split : AP.t A.t -> (Apron.Linexpr1.t * Apron.Linexpr1.t) -> AP.t A.t list

val volume : AP.t A.t -> float

val get_expr : float -> Polka.strict Polka.t A.t -> Apron.Linexpr1.t * Apron.Linexpr1.t

val is_abstraction : AP.t A.t -> Mpqf.t Libabsolute__.Tools.VarMap.t -> bool

given an abstraction and instance, verifies if the abstraction is implied by the instance

val spawn_itv : Apron.Environment.typvar -> Apron.Interval.t -> Libabsolute.Q.t

Random uniform value within an interval, according to the type

val spawn_box : A.box1 -> Libabsolute.Q.t Libabsolute__.Tools.VarMap.t

spawns an instance within a box

val spawner : int -> AP.t A.t -> Libabsolute.Q.t Libabsolute__.Tools.VarMap.t

Takes an integer and compute a spawner function. The integer * corresponds to the number of tries allowed to proceed to the * generation. The bigger it is, the more uniform the spawner will be. A * spawner returns a randomly uniformly chosen instanciation of the * variables. if the polyhedron has a nul (or very small) volume, (e.g * equalities in the constraints) uniformity is not guaranteed

val spawn : AP.t A.t -> Libabsolute.Q.t Libabsolute__.Tools.VarMap.t

val render : AP.t A.t -> Picasso.Drawable.t