Module Libabsolute__.Constraint

type binop =

| ADD

| SUB

| MUL

| DIV

| POW

binary arithmetic operators

type cmpop =

| EQ

| LEQ

| GEQ

| NEQ

| GT

| LT

arithmetic comparison operators

type expr =

| Funcall of string * expr list

| Neg of expr

| Binary of binop * expr * expr

| Var of string

| Cst of Libabsolute__.Q.t

numeric expressions (function call, unary negation, binary operations, variables and constants)

type comparison = expr * cmpop * expr

boolean comparison : e1 <> e2

type 'a boolean =

| Cmp of 'a

| And of 'a boolean * 'a boolean

| Or of 'a boolean * 'a boolean

| Not of 'a boolean

boolean expressions

type t = comparison boolean

type for constraints

val one : expr

Constants

val zero : expr

val two : expr

Expression Constructors

val of_int : int -> expr

builds an expression from an integer

val of_float : float -> expr

builds an expression from an float

val of_mpqf : Libabsolute__.Q.t -> expr

builds an expression from an Mpqf.t

val square : expr -> expr

given an expression e builds the expresspion for e*e

Constructors

val leq : 'a -> 'b -> ('a * cmpop * 'b) boolean

comparisons

val lt : 'a -> 'b -> ('a * cmpop * 'b) boolean

val geq : 'a -> 'b -> ('a * cmpop * 'b) boolean

val gt : 'a -> 'b -> ('a * cmpop * 'b) boolean

val eq : 'a -> 'b -> ('a * cmpop * 'b) boolean

val neq : 'a -> 'b -> ('a * cmpop * 'b) boolean

val assign : string -> Libabsolute__.Q.t -> (expr * cmpop * expr) boolean

constraint for variable assignment by a constant

val inside : 'a -> 'b -> 'b -> ('a * cmpop * 'b) boolean

constraint for 'e \in low;high'

val outside : 'a -> 'b -> 'b -> ('a * cmpop * 'b) boolean

constraint for 'not (v \in low;high)'

val inside_cst : string -> Libabsolute__.Q.t -> Libabsolute__.Q.t -> (expr * cmpop * expr) boolean

same as inside but with constants instead of expressions'

val outside_cst : string -> Libabsolute__.Q.t -> Libabsolute__.Q.t -> (expr * cmpop * expr) boolean

constraint for 'not (v \in low;high)'

Operations

val inv : cmpop -> cmpop

cmp operator inversion

val neg : cmpop -> cmpop

comparison operator negation