Module Libabsolute__.Csp_helper

val one : Libabsolute.Csp.expr

Constants

val zero : Libabsolute.Csp.expr

val two : Libabsolute.Csp.expr

Expression Constructors

val of_int : int -> Libabsolute.Csp.expr

builds an expression from an integer

val of_float : float -> Libabsolute.Csp.expr

builds an expression from an float

val of_mpqf : Libabsolute.Csp.i -> Libabsolute.Csp.expr

builds an expression from an Mpqf.t

val square : Libabsolute.Csp.expr -> Libabsolute.Csp.expr

given an expression e builds the expresspion for e*e

Constraint constructor

val leq : 'a -> 'b -> ('a * Libabsolute.Csp.cmpop * 'b) Libabsolute.Csp.boolean

comparisons

val lt : 'a -> 'b -> ('a * Libabsolute.Csp.cmpop * 'b) Libabsolute.Csp.boolean

val geq : 'a -> 'b -> ('a * Libabsolute.Csp.cmpop * 'b) Libabsolute.Csp.boolean

val gt : 'a -> 'b -> ('a * Libabsolute.Csp.cmpop * 'b) Libabsolute.Csp.boolean

val eq : 'a -> 'b -> ('a * Libabsolute.Csp.cmpop * 'b) Libabsolute.Csp.boolean

val neq : 'a -> 'b -> ('a * Libabsolute.Csp.cmpop * 'b) Libabsolute.Csp.boolean

val assign : Libabsolute.Csp.var -> Libabsolute.Csp.i -> (Libabsolute.Csp.expr * Libabsolute.Csp.cmpop * Libabsolute.Csp.expr) Libabsolute.Csp.boolean

constraint for variable assignment by a constant

val inside : 'a -> 'b -> 'b -> ('a * Libabsolute.Csp.cmpop * 'b) Libabsolute.Csp.boolean

constraint for 'e \in low;high'

val outside : 'a -> 'b -> 'b -> ('a * Libabsolute.Csp.cmpop * 'b) Libabsolute.Csp.boolean

constraint for 'not (v \in low;high)'

val inside_cst : Libabsolute.Csp.var -> Libabsolute.Csp.i -> Libabsolute.Csp.i -> (Libabsolute.Csp.expr * Libabsolute.Csp.cmpop * Libabsolute.Csp.expr) Libabsolute.Csp.boolean

same as inside but with constants instead of expressions'

val outside_cst : Libabsolute.Csp.var -> Libabsolute.Csp.i -> Libabsolute.Csp.i -> (Libabsolute.Csp.expr * Libabsolute.Csp.cmpop * Libabsolute.Csp.expr) Libabsolute.Csp.boolean

constraint for 'not (v \in low;high)'

Predicates

val has_variable : Libabsolute.Csp.expr -> bool

checks if an expression contains a variable

val is_linear : Libabsolute.Csp.expr -> bool

checks if an expression is linear

val is_zero : Mpq.f Mpqf.tt -> bool

val is_neg : 'a Mpqf.tt -> bool

Operations

val inv : Libabsolute.Csp.cmpop -> Libabsolute.Csp.cmpop

cmp operator inversion

val neg : Libabsolute.Csp.cmpop -> Libabsolute.Csp.cmpop

comparison operator negation

val domain_to_constraints : Libabsolute.Csp.decl -> Libabsolute.Csp.bexpr

converts a domain representation to a constraint

val iter_expr : (Libabsolute.Csp.expr -> unit) -> Libabsolute.Csp.expr -> unit

iter on expr

val iter_constr : (Libabsolute.Csp.expr -> unit) -> ((Libabsolute.Csp.expr * 'a * Libabsolute.Csp.expr) Libabsolute.Csp.boolean -> unit) -> (Libabsolute.Csp.expr * 'a * Libabsolute.Csp.expr) Libabsolute.Csp.boolean -> unit

iter on constraints

val map_constr : ('a -> 'b) -> 'a Libabsolute.Csp.boolean -> 'b Libabsolute.Csp.boolean

boolean formulae map

val neg_bexpr : ('a * Libabsolute.Csp.cmpop * 'b) Libabsolute.Csp.boolean -> ('a * Libabsolute.Csp.cmpop * 'b) Libabsolute.Csp.boolean

constraint negation

val remove_not : ('a * Libabsolute.Csp.cmpop * 'b) Libabsolute.Csp.boolean -> ('a * Libabsolute.Csp.cmpop * 'b) Libabsolute.Csp.boolean

rewrites a constraint into an equivalent constraint without 'Not'

val apply : ('a -> 'b -> Libabsolute.Csp.expr * bool) -> 'a -> 'a -> 'b -> Libabsolute.Csp.binop -> Libabsolute.Csp.expr * bool

val distribute : (Libabsolute.Csp.binop * Libabsolute.Csp.expr) -> Libabsolute.Csp.expr -> Libabsolute.Csp.expr

val expand : Libabsolute.Csp.expr -> Libabsolute.Csp.expr

val simplify_expr : Libabsolute.Csp.expr -> bool -> Libabsolute.Csp.expr * bool

val simplify_fp : Libabsolute.Csp.expr -> Libabsolute.Csp.expr

calls simplify until it reaches a fixpoint

val simplify_bexpr : (Libabsolute.Csp.expr * 'a * Libabsolute.Csp.expr) Libabsolute.Csp.boolean -> (Libabsolute.Csp.expr * 'a * Libabsolute.Csp.expr) Libabsolute.Csp.boolean

val left_hand_side : (Libabsolute.Csp.expr * Libabsolute.Csp.cmpop * Libabsolute.Csp.expr) -> Libabsolute.Csp.cmpop * Libabsolute.Csp.expr

val left_hand : (Libabsolute.Csp.expr * Libabsolute.Csp.cmpop * Libabsolute.Csp.expr) Libabsolute.Csp.boolean -> Libabsolute.Csp.cmpop * Libabsolute.Csp.expr

val flatten_pow : Libabsolute.Csp.expr -> int -> Libabsolute.Csp.expr

val derivate : Libabsolute.Csp.expr -> Libabsolute.Csp.var -> Libabsolute.Csp.expr

val derivative : (Libabsolute.Csp.expr * 'a * Libabsolute.Csp.expr) Libabsolute.Csp.boolean -> Libabsolute.Csp.var -> (Libabsolute.Csp.expr * 'a * Libabsolute.Csp.expr) Libabsolute.Csp.boolean

val is_arith : ('a * 'b * 'c) Libabsolute.Csp.boolean -> bool

val ctr_jacobian : (Libabsolute.Csp.expr * Libabsolute.Csp.cmpop * Libabsolute.Csp.expr) Libabsolute.Csp.boolean -> ('a * Libabsolute.Csp.var * 'b) list -> (Libabsolute.Csp.var * Libabsolute.Csp.expr) list

val compute_jacobian : Libabsolute.Csp.problem -> (Libabsolute.Csp.comparison Libabsolute.Csp.boolean * (Libabsolute.Csp.var * Libabsolute.Csp.expr) list) list

val replace_cst_expr : (Libabsolute.Csp.var * Libabsolute.Csp.i) -> Libabsolute.Csp.expr -> Libabsolute.Csp.expr

val replace_cst_bexpr : (Libabsolute.Csp.var * Libabsolute.Csp.i) -> (Libabsolute.Csp.expr * 'a * Libabsolute.Csp.expr) Libabsolute.Csp.boolean -> (Libabsolute.Csp.expr * 'a * Libabsolute.Csp.expr) Libabsolute.Csp.boolean

val get_vars_expr : Libabsolute.Csp.expr -> Libabsolute.Csp.var list

val get_vars_set_expr : Libabsolute.Csp.expr -> Libabsolute__.Tools.VarSet.t

val get_vars_bexpr : (Libabsolute.Csp.expr * 'a * Libabsolute.Csp.expr) Libabsolute.Csp.boolean -> Libabsolute.Csp.var list

val get_vars_set_bexpr : (Libabsolute.Csp.expr * 'a * Libabsolute.Csp.expr) Libabsolute.Csp.boolean -> Libabsolute__.Tools.VarSet.t

val get_vars_jacob : ((Libabsolute.Csp.expr * 'a * Libabsolute.Csp.expr) Libabsolute.Csp.boolean * 'b) list -> ((Libabsolute.Csp.expr * 'a * Libabsolute.Csp.expr) Libabsolute.Csp.boolean * Libabsolute__.Tools.VarSet.t * 'b) list

val replace_cst_jacob : (Libabsolute__.Tools.VarSet.elt * Libabsolute.Csp.i) -> ((Libabsolute.Csp.expr * 'a * Libabsolute.Csp.expr) Libabsolute.Csp.boolean * Libabsolute__.Tools.VarSet.t * 'b) list -> ((Libabsolute.Csp.expr * 'a * Libabsolute.Csp.expr) Libabsolute.Csp.boolean * Libabsolute__.Tools.VarSet.t * 'b) list

val from_cst_to_expr : (Libabsolute.Csp.var * (Libabsolute.Csp.i * Libabsolute.Csp.i)) -> (Libabsolute.Csp.expr * Libabsolute.Csp.cmpop * Libabsolute.Csp.expr) list

val csts_to_expr : (Libabsolute.Csp.var * (Libabsolute.Csp.i * Libabsolute.Csp.i)) list -> (Libabsolute.Csp.expr * Libabsolute.Csp.cmpop * Libabsolute.Csp.expr) list