Module Libabsolute__.Dom

type t =

| Finite of Libabsolute.Q.t * Libabsolute.Q.t
[a;b]

| Minf of Libabsolute.Q.t
]-oo; a]

| Inf of Libabsolute.Q.t
[a; +oo[

| Set of Libabsolute.Q.t list
{x1; x2; ...; xn}

| Top
]-oo; +oo[

main type

Constructors

val interval : Libabsolute.Q.t -> Libabsolute.Q.t -> t

[a;b]

val inf : Libabsolute.Q.t -> t

semi-open interval : [x;+oo[

val set : Libabsolute.Q.t list -> t

finite sets of values {x1; x2; ...; xn}

val minf : Libabsolute.Q.t -> t

semi-open interval : ]-oo;x]

val top : t

]-oo;+oo[

Predicates

val belong : Libabsolute.Q.t -> t -> bool

checks if a given rational belongs to a domain

val is_bounded : t -> bool

Operations

val to_constraint : string -> t -> Libabsolute.Constraint.t

Given a variable name v and a domain d, builds the weakest constraint that v should respect to be in d.

Printing

val print : Stdlib.Format.formatter -> t -> unit

printer

val to_string : t -> string

Conversion to a string