Module Trigo.Make

Parameters

I : Libabsolute__.Itv_sig.ITV

Signature

include I

FILTERING (TEST TRANSFER FUNCTIONS)

include Libabsolute__.Itv_sig.ITV_EVAL

type t

val top_int : t

val top_real : t

val of_ints : int -> int -> t

constructors

val of_rats : Libabsolute.Q.t -> Libabsolute.Q.t -> t

val of_floats : float -> float -> t

val of_int : int -> t

singletons

val of_rat : Libabsolute.Q.t -> t

val of_float : float -> t

val to_float_range : t -> float * float

PRINTING and CONVERSIONS

val to_rational_range : t -> Libabsolute.Q.t * Libabsolute.Q.t

val to_annot : t -> Libabsolute.Csp.typ

returns the type annotation of the represented values

val print : Stdlib.Format.formatter -> t -> unit

val to_constraint : string -> t -> Libabsolute.Constraint.t

converts the interval to a boolean constraint

val float_size : t -> float

measure

val score : t -> float

SET-THEORETIC

val join : t -> t -> t

operations

val meet : t -> t -> t option

val split : t -> t list

val prune : (t -> t -> t list) option

val contains_float : t -> float -> bool

predicates

val is_positive : t -> bool

val is_negative : t -> bool

val is_singleton : t -> bool

val neg : t -> t

INTERVAL ARITHMETICS (FORWARD EVALUATION)

val abs : t -> t

val add : t -> t -> t

val sub : t -> t -> t

val mul : t -> t -> t

val div : t -> t -> t option

return valid values (possibly Bot, if dividend is nul)

val min : t -> t -> t

min and max over intervals

val max : t -> t -> t

val pow : t -> t -> t

returns valid value when the exponant is a singleton positive integer. fails otherwise

val n_root : t -> t -> t option

val exp : t -> t

val ln : t -> t option

val log : t -> t option

val eval_fun : string -> t list -> t option

function calls (sqrt, exp, ln ...) are handled here : given a function name and and a list of argument, it returns a possibly bottom result

val spawn : t -> float

generate a random float within the given interval

val filter_leq : t -> t -> (t * t) Libabsolute.Consistency.t

given two interval arguments, return a filtered subset of each argument by removing points that cannot satisfy the predicate; may also return Sat or Unsat if all (resp. no) point can satisfy the predicate. Interface is simplified since a > b <=> b < a

val filter_lt : t -> t -> (t * t) Libabsolute.Consistency.t

val filter_neq : t -> t -> (t * t) Libabsolute.Consistency.t

val filter_eq : t -> t -> t Libabsolute.Consistency.t

val filter_neg : t -> t -> t option

given the interval argument(s) and the expected interval result of a numeric operation, returns refined interval argument(s) where points that cannot contribute to a value in the result are removed; may also return Bot if no point in an argument can lead to a point in the result

val filter_abs : t -> t -> t option

val filter_add : t -> t -> t -> (t * t) option

val filter_sub : t -> t -> t -> (t * t) option

val filter_mul : t -> t -> t -> (t * t) option

val filter_div : t -> t -> t -> (t * t) option

val filter_pow : t -> t -> t -> (t * t) option

val filter_fun : string -> t list -> t -> t list option

filtering function calls like (sqrt, exp, ln ...) is done here : given a function name, a list of argument, and a result, it remove points that cannot satisfy the relation : f(arg1,..,argn) = r; it returns a possibly bottom result

val pi_down : float

val pi_up : float

val sin_up : float -> float

val sin_down : float -> float

val cos_up : float -> float

val cos_down : float -> float

val tan_up : float -> float

val tan_down : float -> float

val atan_down : float -> float

val atan_up : float -> float

val acos_up : float -> float option

val acos_down : float -> float option

val asin_up : float -> float option

val asin_down : float -> float option

type float_itv = float * float

val print_fitv : Stdlib.Format.formatter -> (float * float) -> unit

val pihalf_down : F.t

val pihalf_up : F.t

val twopiup : float

val twopidown : float

val pihalf_fitv : F.t * F.t

val pi_fitv : F.t * F.t

val twopi_fitv : float * float

val fitv_to_i : (float * float) -> I.t

val pihalf_itv : I.t

val pi_itv : I.t

val twopi_itv : I.t

type monotony =

| Incr

| Decr

| Change of bool

val print_monotony : Stdlib.Format.formatter -> monotony -> unit

val itv : monotony -> ('a -> Stdlib.Float.t) -> ('a -> float) -> ('a * 'a) -> float * float

val cosmonotony : (float * float) -> monotony

val cos_itv : I.t -> I.t

val acos_itv : I.t -> I.t option

val sin_itv : I.t -> I.t

val asin_itv : I.t -> I.t option

val tan_itv : I.t -> I.t option

val atan_itv : I.t -> I.t

val eval_fun : string -> I.t list -> I.t option

val normalize : float -> F.t -> I.t -> I.t * I.t

val arc : I.t -> (I.t -> I.t option) -> I.t -> I.t -> I.t option

val arcos_0_2pi : I.t -> I.t -> I.t option

val arcsin_mpih_pih : I.t -> I.t -> I.t option

val filter : I.t -> (I.t -> I.t -> I.t option) -> I.t -> I.t -> I.t option

val filter_cos : I.t -> I.t -> I.t option

val filter_sin : I.t -> I.t -> I.t option

val arctan_mpih_pih : I.t -> I.t -> I.t option

val filter_tan : I.t -> I.t -> I.t option

val filter_asin : I.t -> I.t -> I.t option

val filter_acos : I.t -> I.t -> I.t option

val filter_atan : I.t -> I.t -> I.t option

val filter_fun : string -> I.t list -> I.t -> I.t list option