Module Libabsolute__ADCP

val complementary : Apron.Environment.t -> Apron.Var.t -> Apron.Scalar.t -> Apron.Linexpr1.t * Apron.Linexpr1.t

module BoxCP : sig ... end

Module for the Box Abstract Domains for Constraint Programming.

module OctMinMinCP : sig ... end

Module for the Octagon Abstract Domains for Constraint Programming.

module OctMinMaxCP : sig ... end

Module for the Octagon Abstract Domains for Constraint Programming.

module OctCP : sig ... end

Module for the Octagon Abstract Domains for Constraint Programming.

module PolyCP : sig ... end

Module for the Polyhedron Abstract Domains for Constraint Programming.