Module Libabsolute__Constraint

This module defines the constraint language, and some basic operations over it

type cmpop =

| EQ

| LEQ

| GEQ

| NEQ

| GT

| LT

arithmetic comparison operators

type comparison = Libabsolute.Expr.t * cmpop * Libabsolute.Expr.t

boolean comparison : e1 <> e2

type 'a boolean =

| Cmp of 'a

| And of 'a boolean * 'a boolean

| Or of 'a boolean * 'a boolean

| Not of 'a boolean

boolean expressions

type t = comparison boolean

type for constraints

Constructors

val leq : Libabsolute.Expr.t -> Libabsolute.Expr.t -> t

Comparisons

e1 <= e2

val lt : Libabsolute.Expr.t -> Libabsolute.Expr.t -> t

e1 < e2

val geq : Libabsolute.Expr.t -> Libabsolute.Expr.t -> t

e1 >= e2

val gt : Libabsolute.Expr.t -> Libabsolute.Expr.t -> t

e1 > e2

val eq : Libabsolute.Expr.t -> Libabsolute.Expr.t -> t

e1 = e2

val neq : Libabsolute.Expr.t -> Libabsolute.Expr.t -> t

e1 <> e2

Boolean formulae

val and_ : t -> t -> t

val or_ : t -> t -> t

val imply : t -> t -> t

val not_ : t -> t

utilities

val assign : string -> Libabsolute.Q.t -> t

constraint for variable assignment by a constant

val inside : Libabsolute.Expr.t -> Libabsolute.Expr.t -> Libabsolute.Expr.t -> t

constraint for 'e \in [low;high]'

val outside : Libabsolute.Expr.t -> Libabsolute.Expr.t -> Libabsolute.Expr.t -> t

constraint for 'e \notin [low;high]'

val inside_cst : string -> Libabsolute.Q.t -> Libabsolute.Q.t -> t

same as inside but with string and rationals instead of expressions'

val outside_cst : string -> Libabsolute.Q.t -> Libabsolute.Q.t -> t

same as outside but with string and rationals instead of expressions'

val of_instance : Libabsolute.Instance.t -> t

conversion of a point p to a conjunctive constraint whose only solution is the point p.

raises [Invalid_arg]

if the instance is empty

val convex_hull : Libabsolute.Instance.t list -> t

Builds the conjuntive constraint defined by the linear system corresponding to the convex hull of the given set of points. The list of instance must not be empty and all the instances should be defined over the same set of keys, otherwise the behaviour in undefined.

Operations

val inv_cmp : cmpop -> cmpop

cmp operator inversion

val neg_cmp : cmpop -> cmpop

comparison operator negation

val cmp_to_fun : cmpop -> Libabsolute.Q.t -> Libabsolute.Q.t -> bool

returns the rational function corresponding to the cmp operator

val nullify_rhs : comparison -> comparison

nullify c computes a comparison equivalent to c, with its right-hand-side being 0

val neg : t -> t

constraint negation

val remove_not : t -> t

rewrites a constraint into an equivalent constraint without the logical operator 'Not'

val collect_vars : t -> int Libabsolute.Tools.VarMap.t

Returns all the variables appearing in a constraint as a map where to each variable is associated the (integer) number of occurences

val replace : t -> string -> Libabsolute.Expr.t -> t

replace constr var expr builds a new constraint identical to constr where all the occurences of the variable var are replaced by the expression expr

val fix_var : t -> string -> Libabsolute.Q.t -> t

fix_var constr var cst builds a new constraint identical to constr where all the occurences of the variable var are replaced by the constant cst

val eval : t -> Libabsolute.Instance.t -> bool

Evaluates the constraint a the given point.

raises [Invalid_arg]

if a division by zero occurs or if an exponentitation by a non integer exposant is made.

Printing

val pp_cmpop : Stdlib.Format.formatter -> cmpop -> unit

comparison operator printer

val pp_comparison : Stdlib.Format.formatter -> comparison -> unit

comparison printer

val print : Stdlib.Format.formatter -> t -> unit

printer

val to_string : t -> string

Conversion to a string