Module Libabsolute__Csp

This module defines the main types used for the constraint language of AbSolute, along with the type of problems, and instance.

Types

type typ =

| Int

| Real

Types of variables

type dom =

| Finite of Libabsolute.Q.t * Libabsolute.Q.t
[a;b]

| Minf of Libabsolute.Q.t
]-oo; a]

| Inf of Libabsolute.Q.t
[a; +oo[

| Set of Libabsolute.Q.t list
{x1; x2; ...; xn}

| Top
]-oo; +oo[

domains of variables

type decl = typ * string * dom

declaration

type instance = Libabsolute.Q.t Libabsolute__.Tools.VarMap.t

the instance type

type info =

| Exact of instance list

| Unfeasible

| Known of (instance * bool) list

annotations to test the validity of the solver

type problem = {

init : decl list;

constraints : Libabsolute.Constraint.t list;

objective : Libabsolute.Expr.t option;

solutions : info option;

}

type of constraint satisfaction problems

Accessors

val get_var_names : problem -> string list

computes the list of all variable names

Constructors

val empty : problem

empty problem, with no variables and no constraints

val initialize : (typ * string * dom) list -> problem

initalizes and unconstrained CSP

val add_real_var : string -> Libabsolute.Q.t -> Libabsolute.Q.t -> problem -> problem

adds a real variable in the csp

val add_real_var_f : string -> float -> float -> problem -> problem

adds a real variable in the csp, with float bounds

val add_int_var : string -> Libabsolute.Q.t -> Libabsolute.Q.t -> problem -> problem

adds an integer variable in the csp

val add_int_var_i : string -> int -> int -> problem -> problem

adds an integer variable in the csp with integer bounds

val add_constr : Libabsolute.Constraint.t -> problem -> problem

adds a constraint to the csp

Printing

val pp_typ : Stdlib.Format.formatter -> typ -> unit

val pp_dom : Stdlib.Format.formatter -> dom -> unit

val pp_declarations : Stdlib.Format.formatter -> (typ * string * dom) list -> unit

val pp_constraints : Stdlib.Format.formatter -> Libabsolute.Constraint.t list -> unit

val pp_instance : Stdlib.Format.formatter -> Libabsolute.Q.t Libabsolute__.Tools.VarMap.t -> unit

val print : Stdlib.Format.formatter -> problem -> unit