Module Libabsolute__Expr

This module defines the numerical language and some basic operations over it

type binop =

| ADD

| SUB

| MUL

| DIV

| POW

binary arithmetic operators

type t =

| Funcall of string * t list

| Neg of t

| Binary of binop * t * t

| Var of string

| Cst of Libabsolute.Q.t

numeric expressions (function call, unary negation, binary operations, variables and constants)

Constructors

val one : t

Constants

val zero : t

val two : t

Expression Constructors

val of_int : int -> t

builds an expression from an integer

val of_float : float -> t

builds an expression from an float

val of_mpqf : Libabsolute.Q.t -> t

builds an expression from an Mpqf.t

val var : string -> t

variables constructor

val add : t -> t -> t

addition

val sub : t -> t -> t

substraction

val mul : t -> t -> t

multiplication

val div : t -> t -> t

multiplication

val square : t -> t

given an expression e builds the expresspion for e*e

Predicates

val has_variable : t -> bool

checks if an expression contains a variable

val is_linear : t -> bool

Operations

val collect_vars : t -> int Libabsolute__.Tools.VarMap.t

Returns all the variables appearing in an expression as a map where to each variable is associated the (integer) number of its occurences

val replace : t -> string -> t -> t

replace expr var value builds a new expression identical to expr where all the occurences of the variable var are replaced by the expression value

val fix_var : t -> string -> Libabsolute.Q.t -> t

fix_var expr var cst builds a new expression identical to expr where all the occurences of the variable var are replaced by the constant cst

val eval : t -> Libabsolute.Instance.t -> Libabsolute.Q.t

Evaluates the expression at the given point.

raises [Invalid_arg]

if a division by zero occurs of if an exponentitation by a non integer exposant is made.

Printing

val pp_var : Stdlib.Format.formatter -> string -> unit

variables printing

val pp_binop : Stdlib.Format.formatter -> binop -> unit

binary operators printing

val print : Stdlib.Format.formatter -> t -> unit

expression printer

val to_string : t -> string

Conversion to a string