Module Libabsolute__Signature

Module signature for Abstract Domains for Constraint Programming (ADCP) they must feature consistency, split and precision operators.

exception TooSmall

can be raised by split operations

module type Numeric = sig ... end

module type Domain = sig ... end

Abstract domain with full handling of boolean expressions instead of only numeric comparisons