Module type Libabsolute__Signature.Solvable

Solvable units. Constraints are internalized

type t

internal representation of a solving state

type space

internal representation of a search space

val propagate : t -> t Libabsolute__.Consistency.t

filtering function

val split : float -> t -> t list

given a float precision p metric and an abstract element, splits the element or raises too small if it is smaller than p

val spawn : t -> Libabsolute.Csp.instance

given an abstract element e returns an arbitrary value in \gamma(e)

val init : Libabsolute.Csp.problem -> t

initalizes a solving state from a csp

val to_result : inner:bool -> space Libabsolute.Result.t -> t -> space Libabsolute.Result.t