Module Libabsolute.Constraint

This module defines the constraint language, and some basic operations over it

type cmpop =

| EQ

| LEQ

| GEQ

| NEQ

| GT

| LT

arithmetic comparison operators

type comparison = Expr.t * cmpop * Expr.t

boolean comparison : e1 <> e2

type 'a boolean =

| Cmp of 'a

| And of 'a boolean * 'a boolean

| Or of 'a boolean * 'a boolean

| Not of 'a boolean

boolean expressions

type t = comparison boolean

type for constraints

val leq : 'a -> 'b -> ('a * cmpop * 'b) boolean

Constructors

val lt : 'a -> 'b -> ('a * cmpop * 'b) boolean

val geq : 'a -> 'b -> ('a * cmpop * 'b) boolean

val gt : 'a -> 'b -> ('a * cmpop * 'b) boolean

val eq : 'a -> 'b -> ('a * cmpop * 'b) boolean

val neq : 'a -> 'b -> ('a * cmpop * 'b) boolean

val assign : string -> Q.t -> (Expr.t * cmpop * Expr.t) boolean

constraint for variable assignment by a constant

val inside : 'a -> 'b -> 'b -> ('a * cmpop * 'b) boolean

constraint for 'e \in low;high'

val outside : 'a -> 'b -> 'b -> ('a * cmpop * 'b) boolean

constraint for 'not (v \in low;high)'

val inside_cst : string -> Q.t -> Q.t -> (Expr.t * cmpop * Expr.t) boolean

same as inside but with constants instead of expressions'

val outside_cst : string -> Q.t -> Q.t -> (Expr.t * cmpop * Expr.t) boolean

constraint for 'not (v \in low;high)'

Operations

val inv : cmpop -> cmpop

cmp operator inversion

val neg : cmpop -> cmpop

comparison operator negation

Printing

val pp_cmpop : Stdlib.Format.formatter -> cmpop -> unit

comparison operator printer

val pp_comparison : Stdlib.Format.formatter -> comparison -> unit

val print : Stdlib.Format.formatter -> t -> unit