Module Libabsolute.Parser

exception Syntax_error of string

exception Semantic_error of string

val constr : (Stdlib.Lexing.lexbuf -> Parser.token) -> Stdlib.Lexing.lexbuf -> Csp.bexpr

Parses a string representation of a constraint. Raises Syntax_error when the input is syntactically invalid

val semantic_check : Csp.problem -> unit

Check that no unbound variables appear, that domains are valid and function calls respect the arity. Raises Semantic_error if one of the above conditions is violated

val file : ?⁠check:bool -> string -> Csp.problem

parses a .abs file and builds the corresponding problem. Raises Syntax_error when the input is syntactically invalid and Semantic_error when the input maakes no sense (e.g unbound variable). When check is set to false, no semantic check is performed