Module Libabsolute__Csp

This module defines the main types used for the constraint language of AbSolute, along with the type of problems, and instance.

Types

type typ =

| Int

| Real

Types of variables

type decl = typ * string * Libabsolute.Dom.t

declaration

type instance = Libabsolute.Q.t Libabsolute.Tools.VarMap.t

the instance type

type info =

| Exact of Libabsolute.Instance.t list

| Unfeasible

| Known of (Libabsolute.Instance.t * bool) list

annotations to test the validity of the solver

type t = {

variables : decl list;

constraints : Libabsolute.Constraint.t list;

objective : Libabsolute.Expr.t option;

solutions : info option;

}

type of constraint satisfaction problems

Accessors

val get_var_names : t -> string list

returns list of all variable names

Constructors

val empty : t

empty problem, with no variables and no constraints

val initialize : (typ * string * Libabsolute.Dom.t) list -> t

initalizes and unconstrained CSP

val add_real_var : string -> Libabsolute.Q.t -> Libabsolute.Q.t -> t -> t

adds a real variable in the csp

val add_real_var_f : string -> float -> float -> t -> t

adds a real variable in the csp, with float bounds

val add_int_var : string -> Libabsolute.Q.t -> Libabsolute.Q.t -> t -> t

adds an integer variable in the csp

val add_int_var_i : string -> int -> int -> t -> t

adds an integer variable in the csp with integer bounds

val add_constr : t -> Libabsolute.Constraint.t -> t

adds a constraint to the csp

Operations

val fix_var : t -> string -> Libabsolute.Q.t -> t

fix_var csp var cst builds a new csp identical to the csp where all the occurences of the variable var are replaced by the constant cst, and where the variables v is removed from the csp variables.

Printing

val pp_typ : Stdlib.Format.formatter -> typ -> unit

val pp_declarations : Stdlib.Format.formatter -> (typ * string * Libabsolute.Dom.t) list -> unit

val pp_constraints : Stdlib.Format.formatter -> Libabsolute.Constraint.t list -> unit

val print : Stdlib.Format.formatter -> t -> unit